Update config location
In recent version of Minio, configuration is
stored in config/config.json, which prevents the
environmental configuration options being stored
in a file called config, since it is not valid to
have a file and directory of the same name in the
same location.

The commit moves the environmental configuration
to the file env, which allows Minio to store its
config in the default location.